requirements.txt
pygame-ce==2.5.5
pyjnius>=1.4.0